+++ Catchy/Catchy/API/Protocol/Course/CourseRepositoryProtocol.swift
@@ -40,8 +40,9 @@ protocol CourseRepositoryProtocol {
     /// 내 코스 조회 API
     func getCourseListData(courseRequest: CourseRequest) -> AnyPublisher<ResponseData<CourseResponse>, MoyaError>
     
-    /// 장소 검색 - 지역명 기반
-    func getPlaceListData(placeSearchRequest: PlaceSearchRequest) -> AnyPublisher<ResponseData<PlaceSearchResponse>, MoyaError>
+    /// 장소 상세 화면 API
+    func getPlaceDetailData(placeId: Int) -> AnyPublisher<ResponseData<PlaceDetailResponse>, MoyaError>
+    
     
 
 }
