+++ Catchy/Catchy/API/Protocol/PlaceCourse/PlaceCourseServiceProtocol.swift
@@ -13,7 +13,7 @@ import Moya
 /// [PlaceCourse] ServiceProtocol
 protocol PlaceCourseServiceProtocol {
     
-    /// 장소 상세 화면 API
-    func getPlaceDetail(placeId: Int) -> AnyPublisher<ResponseData<PlaceDetailResponse>, MoyaError>
+    /// 장소 검색 - 지역명 기반
+    func getPlaceList(placeSearchRequest: PlaceSearchRequest) -> AnyPublisher<ResponseData<PlaceSearchResponse>, MoyaError>
     
 }
