+++ Catchy/Catchy/API/Repository/PlaceCourseRepository.swift
@@ -19,8 +19,8 @@ class PlaceCourseRepository: PlaceCourseRepositoryProtocol {
         self.provider = provider
     }
 
-    /// 장소 검색 - 상세 화면
-    func getPlaceDetailData(placeId: Int) -> AnyPublisher<ResponseData<PlaceDetailResponse>, MoyaError> {
-        return provider.getPlaceDetail(placeId: placeId)
+    /// 장소 검색 - 지역명 기반
+    func getPlaceListData(placeSearchRequest: PlaceSearchRequest) -> AnyPublisher<ResponseData<PlaceSearchResponse>, MoyaError> {
+        return provider.getPlaceList(placeSearchRequest: placeSearchRequest)
     }
 }
