+++ Catchy/Catchy/API/Service/PlaceCourseService.swift
@@ -19,10 +19,10 @@ class PlaceCourseService: PlaceCourseServiceProtocol {
         self.provider = provider
     }
     
-    /// 장소 상세 화면 API
-    func getPlaceDetail(placeId: Int) -> AnyPublisher<ResponseData<PlaceDetailResponse>, MoyaError> {
-        return provider.requestPublisher(.getPlaceDetail(placeId: placeId))
-            .map(ResponseData<PlaceDetailResponse>.self)
+    /// 장소 검색 - 지역명 기반
+    func getPlaceList(placeSearchRequest: PlaceSearchRequest) -> AnyPublisher<ResponseData<PlaceSearchResponse>, MoyaError> {
+        return provider.requestPublisher(.getPlaceList(place: placeSearchRequest))
+            .map(ResponseData<PlaceSearchResponse>.self)
             .eraseToAnyPublisher()
     }
 }
