--- Catchy/Catchy/Model/Request/GetCourse/GetCourseRequest.swift
@@ -1,15 +1,0 @@
-//
-//  GetCourseRequest.swift
-//  Catchy
-//
-//  Created by 정의찬 on 1/23/25.
-//
-
-import Foundation
-
-struct GetCourseRequest: Codable {
-    let type: CourseType
-    let upperLocation: String
-    let lowerLocation: String
-    let lastId: Int
-}
